Move upload logic to new file
